Added what I think is Category from Jira

+++ keyinsight/src/main/java/com/westerndigital/keyinsight/JavaProject/JavaProjectRepository.java
@@ -2,6 +2,6 @@ package com.westerndigital.keyinsight.JavaProject;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JavaProjectRepository extends JpaRepository<JavaProject, Long> {
+public interface JavaProjectRepository extends JpaRepository<JavaProject, Integer> {
 
 }
